:art: Move test output to tempdir
