+++ .gitignore
@@ -1,1 +1,2 @@
-/bazel-*
+*.pyc
+.cache
